Move to a 64-bit timestamp
And to a 63-bit data-size, with unlimited channels

+++ SimBusCommon/Bus/SBEvent.h
@@ -29,17 +29,17 @@ typedef enum
 |* Set the designated initialiser
 \*****************************************************************************/
 - (instancetype) init NS_UNAVAILABLE;
-- (instancetype) initWithType:(EventType)type at:(uint32_t)when;
+- (instancetype) initWithType:(EventType)type at:(uint64_t)when;
 
 /*****************************************************************************\
 |* Convenience initialisers
 \*****************************************************************************/
-+ (instancetype) withAbsoluteTime:(uint32_t)ns;
-+ (instancetype) withRelativeTime:(uint32_t)ns;
-+ (instancetype) afterNextClockHi:(uint32_t)ns;
-+ (instancetype) afterNextClockLo:(uint32_t)ns;
-+ (instancetype) beforeNextClockHi:(uint32_t)ns;
-+ (instancetype) beforeNextClockLo:(uint32_t)ns;
++ (instancetype) withAbsoluteTime:(uint64_t)ns;
++ (instancetype) withRelativeTime:(uint64_t)ns;
++ (instancetype) afterNextClockHi:(uint64_t)ns;
++ (instancetype) afterNextClockLo:(uint64_t)ns;
++ (instancetype) beforeNextClockHi:(uint64_t)ns;
++ (instancetype) beforeNextClockLo:(uint64_t)ns;
 + (instancetype) onSignalChange:(NSArray<NSString *> *)signals;
 
 /*****************************************************************************\
@@ -50,7 +50,7 @@ typedef enum
 @property(assign, nonatomic) EventType                              type;
 
 // The time-value of the event, in nanoseconds
-@property(assign, nonatomic) uint32_t                               when;
+@property(assign, nonatomic) uint64_t                               when;
 
 // The plugin that should process this event
 @property(strong, nonatomic) id<SBPlugin>                           plugin;
