+++ SimBusCommon/Bus/SBEvent.m
@@ -12,7 +12,7 @@
 /*****************************************************************************\
 |* Initialise an instance
 \*****************************************************************************/
-- (instancetype) initWithType:(EventType)type at:(uint32_t)when
+- (instancetype) initWithType:(EventType)type at:(uint64_t)when
     {
     if (self = [super init])
         {
@@ -26,32 +26,32 @@
 /*****************************************************************************\
 |* Convenience initialisers
 \*****************************************************************************/
-+ (instancetype) withAbsoluteTime:(uint32_t)ns
++ (instancetype) withAbsoluteTime:(uint64_t)ns
     {
     return [[SBEvent alloc] initWithType:AbsoluteTimeEvent at:ns];
     }
     
-+ (instancetype) withRelativeTime:(uint32_t)ns
++ (instancetype) withRelativeTime:(uint64_t)ns
     {
     return [[SBEvent alloc] initWithType:RelativeTimeEvent at:ns];
     }
 
-+ (instancetype) afterNextClockHi:(uint32_t)ns
++ (instancetype) afterNextClockHi:(uint64_t)ns
     {
     return [[SBEvent alloc] initWithType:AfterNextClockHiEvent at:ns];
     }
 
-+ (instancetype) afterNextClockLo:(uint32_t)ns
++ (instancetype) afterNextClockLo:(uint64_t)ns
     {
     return [[SBEvent alloc] initWithType:AfterNextClockLoEvent at:ns];
     }
 
-+ (instancetype) beforeNextClockHi:(uint32_t)ns
++ (instancetype) beforeNextClockHi:(uint64_t)ns
     {
     return [[SBEvent alloc] initWithType:BeforeNextClockHiEvent at:ns];
     }
 
-+ (instancetype) beforeNextClockLo:(uint32_t)ns
++ (instancetype) beforeNextClockLo:(uint64_t)ns
     {
     return [[SBEvent alloc] initWithType:BeforeNextClockLoEvent at:ns];
     }
